validator stats with context and hooks
