+++ src/ValidatorStatsContext.js
@@ -1,0 +1,5 @@
+import React from "react"
+
+const ValidatorStatsContext = React.createContext()
+
+export default ValidatorStatsContext
